% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering2df90.R
\name{clusteringf90}
\alias{clusteringf90}
\title{this is a wrapper function for the fortran subroutine clustering}
\usage{
clusteringf90(x, startID = 1, thres = 0)
}
\description{
this is a wrapper function for the fortran subroutine clustering
}

