% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringR.R
\name{clusteringR}
\alias{clusteringR}
\title{detect clusters in a matrix}
\usage{
clusteringR(x, thres = 0)
}
\arguments{
\item{x}{the data matrix}

\item{thres}{threshold}
}
\description{
detect clusters in a matrix
}

